Clippy fixes

+++ imessage-exporter/src/app/runtime.rs
@@ -97,10 +97,10 @@ impl Config {
 
     /// Get the attachment path for a specific chat ID
     pub fn conversation_attachment_path(&self, chat_id: Option<i32>) -> String {
-        if let Some(chat_id) = chat_id {
-            if let Some(real_id) = self.real_chatrooms.get(&chat_id) {
-                return real_id.to_string();
-            }
+        if let Some(chat_id) = chat_id
+            && let Some(real_id) = self.real_chatrooms.get(&chat_id)
+        {
+            return real_id.to_string();
         }
         String::from(ORPHANED)
     }
@@ -456,12 +456,12 @@ impl Config {
             self.run_diagnostic()?;
         } else if let Some(export_type) = &self.options.export_type {
             // Ensure that if we want to filter on things, we have stuff to filter for
-            if let Some(filters) = &self.options.conversation_filter {
-                if !self.options.query_context.has_filters() {
-                    return Err(RuntimeError::InvalidOptions(format!(
-                        "Selected filter `{filters}` does not match any participants!"
-                    )));
-                }
+            if let Some(filters) = &self.options.conversation_filter
+                && !self.options.query_context.has_filters()
+            {
+                return Err(RuntimeError::InvalidOptions(format!(
+                    "Selected filter `{filters}` does not match any participants!"
+                )));
             }
 
             // Ensure the path we want to export to exists
@@ -595,17 +595,17 @@ impl Drop for Config {
     fn drop(&mut self) {
         if let Some(backup) = &self.backup {
             // Remove the temporary `sms.db` file if it was created
-            if backup.manifest_db.is_temporary {
-                if let Some(conn) = self.db.take() {
-                    let path = conn.path().unwrap().to_string();
-                    conn.close().ok();
+            if backup.manifest_db.is_temporary
+                && let Some(conn) = self.db.take()
+            {
+                let path = conn.path().unwrap().to_string();
+                conn.close().ok();
 
-                    // Remove the file, ignoring errors if any
-                    if let Err(e) = remove_file(&path) {
-                        eprintln!(
-                            "warning: failed to remove temporary messages database at {path}: {e}"
-                        );
-                    }
+                // Remove the file, ignoring errors if any
+                if let Err(e) = remove_file(&path) {
+                    eprintln!(
+                        "warning: failed to remove temporary messages database at {path}: {e}"
+                    );
                 }
             }
         }
